aggiunta service e controller di prova
separata la logica della miniera aggiungendo minieraService e MiningResultDTO

aggiunta di classe Shop (ancora senza db)

aggiunta controller di base per testare l'interfaccia

+++ src/main/java/it/unicam/cs/mpgc/rpg125936/domain/user/Player.java
@@ -7,24 +7,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Player extends User{
-    private int money;
+    private double money;
     // private int pickaxeDurability;
     // Lista di oggetti (armi, materiali, etc.)
 
-    public Player(String name, int initialMoney) {
+    public Player(String name, double initialMoney) {
         super(name);
         this.money = initialMoney;
     }
 
-    // Metodi per gestire la salute, i soldi e lo scavo
-    public void addMoney(int amount) { this.money += amount; }
-    //public void reducePickaxe() { this.pickaxeDurability--; }
+    public void addMoney(double amount) { this.money += amount; }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
